+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 * Change streching in some fields of main window (#599)
 * Accept floats for lua draw commands
 * Don't sleep on main thread when fast-forwarding
+* Don't execute lua onPaint callbacks when non rendering to improve fast-forward
 
 ### Fixed